backend/api/requirements.txt
flask
mysql-connector-python
flask_cors
requests
uuid
dotenv
twilio
flask_mail